+++ src/roq/samples/event-log/application.h
@@ -1,0 +1,26 @@
+/* Copyright (c) 2017-2021, Hans Erik Thrane */
+
+#pragma once
+
+#include "roq/span.h"
+#include "roq/tool.h"
+
+namespace roq {
+namespace samples {
+namespace event_log {
+
+// Service is a base class used to initialize the environment
+// it is not required, only for your convenience
+
+class Application final : public Tool {
+ public:
+  using Tool::Tool;
+
+ protected:
+  int main_helper(const roq::span<std::string_view> &args);
+  int main(int argc, char **argv) override;
+};
+
+}  // namespace event_log
+}  // namespace samples
+}  // namespace roq